Moved some << operators to tests, deleted others.
