+++ include/terminalpp/element.hpp
@@ -55,10 +55,4 @@ constexpr bool operator!=(element const &lhs, element const &rhs)
     return !(lhs == rhs);
 }
 
-//* =========================================================================
-/// \brief Output stream operator
-//* =========================================================================
-TERMINALPP_EXPORT
-std::ostream &operator<<(std::ostream &out, element const &element);
-
 }
